Move GUI from examples/cra to packages/gui
