+++ examples/cra/package.json
@@ -14,6 +14,7 @@
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-scripts": "3.1.1",
+    "tailwindcss": "^1.1.2",
     "typescript": "^3.5.3"
   },
   "scripts": {
