+++ examples/cra/public/index.html
@@ -21,7 +21,6 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
   <title>Tailwind CSS GUI</title>
-  <link crossorigin="anonymous" rel="stylesheet" href="https://unpkg.com/tailwindcss@1.1.0/dist/tailwind.min.css">
 </head>
 
 <body>
