+++ packages/gui/package.json
@@ -5,7 +5,8 @@
   "dependencies": {
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
-    "react-scripts": "3.1.1"
+    "react-scripts": "3.1.1",
+    "tailwindcss": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
