+++ packages/gui/src/components/Inspector/Rule.tsx
