+++ packages/gui/src/components/Inspector/index.tsx
