+++ packages/gui/src/components/Inspector/searchMeta.ts
