+++ packages/gui/src/components/Inspector/useSearch.ts
