+++ packages/gui/src/components/Panel/index.tsx
